Subsampling of partitions to generic pipeline
