+++ heareval/tasks/speech_commands.py
@@ -57,6 +57,11 @@ config = {
         "test": "http://download.tensorflow.org/data/speech_commands_test_set_v0.02.tar.gz",  # noqa: E501
     },
     "sample_duation": 1.0,
+    "partitions": [
+        {"name": "train", "max_files": 100},
+        {"name": "test", "max_files": 100},
+        {"name": "valid", "max_files": 100},
+    ],
 }
 
 
@@ -228,102 +233,102 @@ class ConfigureProcessMetaData(luigi_util.WorkTask):
         self.mark_complete()
 
 
-class SubsamplePartition(luigi_util.SubsamplePartition):
-    """
-    A subsampler that acts on a specific partition.
-    All instances of this will depend on the combined process metadata csv.
-    """
-
-    def requires(self):
-        # The meta files contain the path of the files in the data
-        # so we dont need to pass the extract as a dependency here.
-        return {
-            "meta": ConfigureProcessMetaData(outfile="process_metadata.csv"),
-        }
-
-
-class SubsamplePartitions(luigi_util.WorkTask):
-    """
-    Aggregates subsampling of all the partitions into a single task as dependencies.
-    All the subsampled files are stored in the requires workdir, so we just link to
-    that since there aren't any real outputs associated with this task.
-    This is a bit of a hack -- but it allows us to avoid rewriting
-    the Subsample task as well as take advantage of Luigi concurrency.
-    """
-
-    def requires(self):
-        # Perform subsampling on each partition independently
-        subsample_partitions = {
-            partition.name: SubsamplePartition(
-                partition=partition.name, max_files=partition.max_files
-            )
-            for partition in config.partitions
-        }
-        return subsample_partitions
-
-    def run(self):
-        workdir = Path(self.workdir)
-        workdir.rmdir()
-        workdir.symlink_to(Path(self.requires()["train"].workdir).absolute())
-        self.mark_complete()
-
-
-class MonoWavTrimCorpus(luigi_util.MonoWavTrimCorpus):
-    def requires(self):
-        return {"corpus": SubsamplePartitions()}
-
-
-class SplitTrainTestCorpus(luigi_util.SplitTrainTestCorpus):
-    def requires(self):
-        # The metadata helps in provide the partition type for each
-        # audio file
-        return {
-            "corpus": MonoWavTrimCorpus(duration=config.sample_duration),
-            "meta": ConfigureProcessMetaData(outfile="process_metadata.csv"),
-        }
-
-
-class SplitTrainTestMetadata(luigi_util.SplitTrainTestMetadata):
-    def requires(self):
-        # Requires the traintestcorpus and the metadata.
-        # The metadata is split into train and test files
-        # which are in the traintestcorpus
-        return {
-            "traintestcorpus": SplitTrainTestCorpus(),
-            "meta": ConfigureProcessMetaData(outfile="process_metadata.csv"),
-        }
-
-
-class MetadataVocabulary(luigi_util.MetadataVocabulary):
-    def requires(self):
-        # Depends only on the train test metadata
-        return {"traintestmeta": SplitTrainTestMetadata()}
-
-
-class ResampleSubCorpus(luigi_util.ResampleSubCorpus):
-    def requires(self):
-        # Requires the train test corpus and will take in
-        # parameter for which partition and sr the resampling
-        # has to be done
-        return {"traintestcorpus": SplitTrainTestCorpus()}
-
-
-class FinalizeCorpus(luigi_util.WorkTask):
-
-    sample_rates = luigi.ListParameter()
-    next_task = luigi.TaskParameter()
-
-    def requires(self):
-        # Will copy the resampled data and the traintestmeta and the vocabmeta
-        return {
-            "resample": [
-                ResampleSubCorpus(sr=sr, partition=partition)
-                for sr in self.sample_rates
-                for partition in ["train", "test", "valid"]
-            ],
-            "traintestmeta": SplitTrainTestMetadata(),
-            "vocabmeta": MetadataVocabulary(),
-        }
+# class SubsamplePartition(luigi_util.SubsamplePartition):
+#     """
+#     A subsampler that acts on a specific partition.
+#     All instances of this will depend on the combined process metadata csv.
+#     """
+#
+#     def requires(self):
+#         # The meta files contain the path of the files in the data
+#         # so we dont need to pass the extract as a dependency here.
+#         return {
+#             "meta": ConfigureProcessMetaData(outfile="process_metadata.csv"),
+#         }
+#
+#
+# class SubsamplePartitions(luigi_util.WorkTask):
+#     """
+#     Aggregates subsampling of all the partitions into a single task as dependencies.
+#     All the subsampled files are stored in the requires workdir, so we just link to
+#     that since there aren't any real outputs associated with this task.
+#     This is a bit of a hack -- but it allows us to avoid rewriting
+#     the Subsample task as well as take advantage of Luigi concurrency.
+#     """
+#
+#     def requires(self):
+#         # Perform subsampling on each partition independently
+#         subsample_partitions = {
+#             partition.name: SubsamplePartition(
+#                 partition=partition.name, max_files=partition.max_files
+#             )
+#             for partition in config.partitions
+#         }
+#         return subsample_partitions
+#
+#     def run(self):
+#         workdir = Path(self.workdir)
+#         workdir.rmdir()
+#         workdir.symlink_to(Path(self.requires()["train"].workdir).absolute())
+#         self.mark_complete()
+#
+#
+# class MonoWavTrimCorpus(luigi_util.MonoWavTrimCorpus):
+#     def requires(self):
+#         return {"corpus": SubsamplePartitions()}
+#
+#
+# class SplitTrainTestCorpus(luigi_util.SplitTrainTestCorpus):
+#     def requires(self):
+#         # The metadata helps in provide the partition type for each
+#         # audio file
+#         return {
+#             "corpus": MonoWavTrimCorpus(duration=config.sample_duration),
+#             "meta": ConfigureProcessMetaData(outfile="process_metadata.csv"),
+#         }
+#
+#
+# class SplitTrainTestMetadata(luigi_util.SplitTrainTestMetadata):
+#     def requires(self):
+#         # Requires the traintestcorpus and the metadata.
+#         # The metadata is split into train and test files
+#         # which are in the traintestcorpus
+#         return {
+#             "traintestcorpus": SplitTrainTestCorpus(),
+#             "meta": ConfigureProcessMetaData(outfile="process_metadata.csv"),
+#         }
+#
+#
+# class MetadataVocabulary(luigi_util.MetadataVocabulary):
+#     def requires(self):
+#         # Depends only on the train test metadata
+#         return {"traintestmeta": SplitTrainTestMetadata()}
+#
+#
+# class ResampleSubCorpus(luigi_util.ResampleSubCorpus):
+#     def requires(self):
+#         # Requires the train test corpus and will take in
+#         # parameter for which partition and sr the resampling
+#         # has to be done
+#         return {"traintestcorpus": SplitTrainTestCorpus()}
+#
+#
+# class FinalizeCorpus(luigi_util.WorkTask):
+#
+#     sample_rates = luigi.ListParameter()
+#     next_task = luigi.TaskParameter()
+#
+#     def requires(self):
+#         # Will copy the resampled data and the traintestmeta and the vocabmeta
+#         return {
+#             "resample": [
+#                 ResampleSubCorpus(sr=sr, partition=partition)
+#                 for sr in self.sample_rates
+#                 for partition in ["train", "test", "valid"]
+#             ],
+#             "traintestmeta": SplitTrainTestMetadata(),
+#             "vocabmeta": MetadataVocabulary(),
+#         }
 
 
 def main(num_workers: int, sample_rates: List[int]):
@@ -341,8 +346,12 @@ def main(num_workers: int, sample_rates: List[int]):
         data_config=config,
     )
 
+    final = pipelines.SubsamplePartitions(
+        metadata=configure_metadata, data_config=config
+    )
+
     luigi.build(
-        [configure_metadata],
+        [final],
         # [FinalizeCorpus(sample_rates=sample_rates, task_string=config.versioned_task_name, next_task=GenerateTrainDataset)],
         workers=num_workers,
         local_scheduler=True,
